RF: Move peaks and related tests to dipy.direction. Adjust imports.
