+++ dipy/direction/__init__.py
@@ -1,4 +1,4 @@
 
 from .probabilistic_direction_getter import (ProbabilisticDirectionGetter,
                                              DeterministicMaximumDirectionGetter)
-from dipy.reconst.peaks import *
+import peaks
