+++ dipy/direction/peaks.py
@@ -12,12 +12,13 @@ from nibabel.tmpdirs import InTemporaryDirectory
 import numpy as np
 import scipy.optimize as opt
 
-from .recspeed import local_maxima, remove_similar_vertices, search_descending
+from dipy.reconst.recspeed import (local_maxima, remove_similar_vertices,
+                                   search_descending)
 from dipy.core.sphere import HemiSphere, Sphere
 from dipy.data import default_sphere
 from dipy.core.ndindex import ndindex
 from dipy.reconst.shm import sh_to_sf_matrix
-from .peak_direction_getter import PeaksAndMetricsDirectionGetter
+from dipy.reconst.peak_direction_getter import PeaksAndMetricsDirectionGetter
 
 
 def peak_directions_nl(sphere_eval, relative_peak_threshold=.25,
