+++ dipy/direction/probabilistic_direction_getter.py
@@ -1,7 +1,8 @@
-"""Implementation of a probabilistic direction getter based on sampling from
+"""
+Implementation of a probabilistic direction getter based on sampling from
 discrete distribution (pmf) at each step of the tracking."""
 import numpy as np
-from dipy.reconst.peaks import peak_directions, default_sphere
+from .peaks import peak_directions, default_sphere
 from dipy.reconst.shm import order_from_ncoef, sph_harm_lookup
 from dipy.tracking.local.direction_getter import DirectionGetter
 from dipy.tracking.local.interpolation import trilinear_interpolate4d
@@ -101,7 +102,7 @@ class ProbabilisticDirectionGetter(PeakDirectionGetter):
 
         See also
         --------
-        dipy.reconst.peaks.peak_directions
+        dipy.direction.peaks.peak_directions
 
 
         """
@@ -147,7 +148,7 @@ class ProbabilisticDirectionGetter(PeakDirectionGetter):
 
         See also
         --------
-        dipy.reconst.peaks.peak_directions
+        dipy.direction.peaks.peak_directions
 
         """
         pmf_gen = SHCoeffPmfGen(shcoeff, sphere, basis_type)
@@ -175,7 +176,7 @@ class ProbabilisticDirectionGetter(PeakDirectionGetter):
 
         See also
         --------
-        dipy.reconst.peaks.peak_directions
+        dipy.direction.peaks.peak_directions
 
         """
         PeakDirectionGetter.__init__(self, sphere, **kwargs)
