+++ dipy/reconst/tests/test_mapmri_odf.py
@@ -1,7 +1,7 @@
 import numpy as np
 from dipy.data import get_sphere, get_3shell_gtab, get_isbi2013_2shell_gtab
 from dipy.reconst.mapmri import MapmriModel
-from dipy.reconst.peaks import gfa, peak_directions
+from dipy.direction.peaks import gfa, peak_directions
 from numpy.testing import (assert_equal,
                            assert_almost_equal,
                            run_module_suite,
@@ -53,7 +53,7 @@ def test_mapmri_odf():
             assert_equal(gfa(odf) < 0.1, True)
 
 
-def test_multivox_mapmri():    
+def test_multivox_mapmri():
     gtab = get_3shell_gtab()
 
     data = np.random.random([20, 30, 1, gtab.gradients.shape[0]])
